# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: RAG platform public API
  version: v1
  description: Multi-tenant company-knowledge RAG platform. Tenant is derived from the authenticated principal (FR-ACC-03); there is no tenant_id parameter on tenant-scoped routes (SPEC-07 §1).
servers:
  - url: /
    description: Same-origin; base path /v1 (SPEC-07 §1).
...
paths:
...
  /v1/retrieve:
    post:
      summary: 'Hybrid retrieval: embed the query and return ranked chunks with scores and citation metadata (no generation). score is the fused RRF score, or the reranker score when the tenant enables reranking. Body: {query, top_k?, filters?}.'
      operationId: retrieve
      tags:
        - retrieval
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ranked chunks
        "400":
          description: missing query or malformed body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "401":
          description: missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: API key lacks the required scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "429":
          description: rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "503":
          description: tenant is not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
    ErrorEnvelope:
      additionalProperties: false
      properties:
        error:
          additionalProperties: false
          properties:
            code:
              enum:
                - unauthorized
                - forbidden
                - not_found
                - validation
                - rate_limited
                - tenant_unavailable
                - conflict
                - internal
              type: string
            message:
              type: string
            request_id:
              type: string
          required:
            - code
            - message
          type: object
      required:
        - error
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'API key: Authorization: Bearer rk_<prefix>_<secret> (SPEC-07 §2, ADR-0021). Tenant is derived from the key.'
      bearerFormat: rk_<hexprefix>_<secret>
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Server-side session cookie (SPEC-09 §3). Mutations also require the X-CSRF-Token header.